% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load farse data from csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A csv file name that will be imported into tibble}
}
\value{
This function will return tibble from imported csv
}
\description{
This function will used to import csv file to tydeverse tibble format.
 Import will check if file exists. In case that file does not exist, function will stop.
}
