% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Function fars_read_years reads in multiple files and creates list of tibbls}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Requires vector of years, that will be imported into list of tibbles}
}
\value{
Function will return list of tibbles
}
\description{
Function will take in vector of years, either in string or numbers.
 fars_read_years will use functions make_filename, to create file name
 to by loaded into list. Function fars_read will be also used in order to
 read in cvs files. Final step of the function will be adding new column year
 from variable year.
}
\details{
fars_read_years will have check on existance of csv file with year, that was entered
 in vector of years.
}
\examples{

fars_read_years(c(2013, 2014))
fars_read_years(c('2013', '2014'))

}
