% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function fars_summarize_years creates summarized version of years and months}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character string giving the text the function will print}
}
\value{
This function will return tibble of values that have number of
         combinations of months and years.
}
\description{
This function uses output from fars_read_years which has output of list of
 tibbles. Secund step is creating union of all tibbles into one table. Third step
 is creating summaries of number combinations by year and month. final step is
 to create new table with values of month in columns and number of existing
 combinations as values in columns.
}
