% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create name of file from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{}
}
\value{
Returns string that contains file name
}
\description{
Simple function, that will create file name from imput string or
 number. In case that input is not numeric, than function changes
 type to numeric.
}
\examples{

make_filename('2015')
make_filename(2015)

}
