% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storm_size.R
\name{plot_storm_size}
\alias{plot_storm_size}
\title{A function to plot a map of the position and size of a given storm
It would plot based on 34, 50, and 64 knot extent variable}
\usage{
plot_storm_size(row, df)
}
\arguments{
\item{row}{| the strongest landfall determined by maximum sustained wind speed of the specified hurricane}

\item{df}{| dataframe that contains all the specified hurricane with name and year}
}
\value{
p | ggplot object
}
\description{
A function to plot a map of the position and size of a given storm
It would plot based on 34, 50, and 64 knot extent variable
}
\examples{
data(processed_hurdat)
katrina = plot_storm_size_helper(processed_hurdat, "katrina", "2005")
p1 = plot_storm_size(katrina[1,],katrina[-1,])
}
