% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storm_tracks.R
\name{plot_storm_tracks}
\alias{plot_storm_tracks}
\title{Plot storm track on the map}
\usage{
plot_storm_tracks(storm_data, storm_ids)
}
\arguments{
\item{storm_data}{| the dataframe for the hurricane information}

\item{storm_ids}{| the storm IDs to be plot for its tracks}
}
\value{
p | a ggplot object
}
\description{
Plot storm track on the map
}
\examples{
data(processed_hurdat)
library(mapdata)
library(ggplot2)
library(maps)
p <- plot_storm_tracks(processed_hurdat,c("AL022018","AL042017"))
}
